dinsy-server/requirements.txt
configargparse-0.12.0
python-mimeparse-1.6.0
flask-0.12